refactor: move end/start_song to socket, move nowplaying to blueprint
